cleaned up middleware functions
