+++ backend/src/middleware/auth/index.js
@@ -30,23 +30,3 @@ export const checkToken = (req, res, next) => {
     }
   });
 };
-
-// Make sure that the authId matches the userId that is
-// being fetched via the userId path param
-export const checkUser = (req, res, next) => {
-  // This is made available via the firebase checkToken middleware
-  const { authId } = req;
-  const userId = req.params.userId;
-
-  if (authId == userId) {
-    return next();
-  } else {
-    return handleApiError(res, new AuthenticationError());
-  }
-};
-
-// Check if the user is of type teacher
-export const checkTeacherRole = (req, res, next) => {
-  // TODO: Implement this when we have a solution for user roles
-  return next();
-};
