+++ backend/src/middleware/roles/index.js
@@ -1,0 +1,5 @@
+// Check if the user is of type teacher
+export const checkTeacherRole = (req, res, next) => {
+  // TODO: Implement this when we have a solution for user roles
+  return next();
+};
